refactor: us-0002 update followers count

+++ src/main/java/com/mercadolibre/be_java_hisp_w31_g3/dto/UserDto.java
@@ -17,11 +17,11 @@ public class UserDto {
     @Builder.Default
     private List<UserDto> followed = null;
     @Builder.Default
-    private Integer followersCount = null;
+    private Long followersCount = null;
 
     public UserDto(){}
 
-    public UserDto(Long userId, String userName, List<UserDto> followers, List<UserDto> followed, Integer followersCount) {
+    public UserDto(Long userId, String userName, List<UserDto> followers, List<UserDto> followed, Long followersCount) {
         this.userId = userId;
         this.userName = userName;
         this.followers = followers;
